+++ vss_cli/const.py
@@ -5,7 +5,7 @@ import pkg_resources
 
 PACKAGE_NAME = "vss_cli"
 
-__version__ = "0.1.9"
+__version__ = "0.2.0"
 
 REQUIRED_PYTHON_VER = (3, 6, 4)
 
@@ -251,10 +251,19 @@ COLUMNS_MESSAGE = [
     ("SUBJECT", "subject"),
     ("TEXT", "text"),
 ]
+COLUMNS_VM_TEMPLATE = [
+    *COLUMNS_VM_MIN,
+    ("FOLDER", "folder.path"),
+    ("CPU", "cpuCount"),
+    ("MEMORY", "memoryGB"),
+    ("GUEST", "guestFullName"),
+    ("VERSION", "version"),
+]
 COLUMNS_VM = [
     *COLUMNS_VM_MIN,
-    ("FOLDER", "folder"),
+    ("FOLDER", "folder.path"),
     ("CPU", "cpuCount"),
+    ("IP_ADDRESS", "ipAddress"),
     ("MEMORY", "memoryGB"),
     ("POWER", "powerState"),
     ("GUEST", "guestFullName"),
@@ -474,3 +483,13 @@ COLUMNS_SSH_KEY = [
     ("FINGERPRINT", "fingerprint"),
     ("VALUE", "value"),
 ]
+
+VM_DISK_MODES = [
+    'persistent',
+    'nonpersistent',
+    'undoable',
+    'independent_persistent',
+    'independent_nonpersistent',
+    'append',
+]
+VM_SCSI_TYPES = ['paravirtual', 'lsilogic', 'lsilogicsas', 'buslogic']